Update topics EE to use tabs provider
