+++ packages/extension/src/view/devtools/components/privateAdvertising/topics/index.tsx
@@ -17,23 +17,21 @@
 /**
  * External dependencies.
  */
-import React, { useMemo, useState } from 'react';
+import React, { useMemo } from 'react';
 import {
   InfoCard as InfoCardTemplate,
   PSInfoKey,
   QuickLinksList,
-  Tabs,
+  TabsProvider,
   type PSInfoKeyType,
   type TabItems,
 } from '@google-psat/design-system';
-import { I18n } from '@google-psat/i18n';
-import classNames from 'classnames';
-
 /**
  * Internal dependencies.
  */
 import ExplorableExplanation from './explorableExplanation';
 import TaxonomyTree from './taxonomyTree';
+import Panel from './panel';
 
 const InfoCard = ({ infoKey }: { infoKey: PSInfoKeyType }) => {
   return (
@@ -47,8 +45,6 @@ const InfoCard = ({ infoKey }: { infoKey: PSInfoKeyType }) => {
 };
 
 const Topics = () => {
-  const [activeTab, setActiveTab] = useState(0);
-
   const tabItems = useMemo<TabItems>(
     () => [
       {
@@ -97,37 +93,10 @@ const Topics = () => {
     []
   );
 
-  const ActiveTabContent = tabItems[activeTab].content.Element;
-
   return (
-    <div
-      data-testid="topics-content"
-      className="h-screen w-full flex flex-col overflow-hidden"
-    >
-      <div className="p-4">
-        <div className="flex gap-2 text-2xl font-bold items-baseline text-raisin-black dark:text-bright-gray">
-          <h1 className="text-left">{I18n.getMessage('topics')}</h1>
-        </div>
-      </div>
-      <Tabs
-        items={tabItems}
-        activeTab={activeTab}
-        setActiveTab={setActiveTab}
-      />
-      <div
-        className={classNames(
-          'overflow-auto flex-1',
-          tabItems[activeTab].content.className
-        )}
-        style={{
-          minHeight: 'calc(100% - 93px)',
-        }}
-      >
-        {ActiveTabContent && (
-          <ActiveTabContent {...tabItems[activeTab].content.props} />
-        )}
-      </div>
-    </div>
+    <TabsProvider items={tabItems}>
+      <Panel />
+    </TabsProvider>
   );
 };
